+++ packages/abi-coder/src/interface.test.ts
@@ -82,7 +82,7 @@ describe('Interface', () => {
       jsonFragmentTwoParams,
       jsonFragmentTwoParams,
     ]);
-    expect(fragmentTwoParams.format()).toBe('sum_test(s(u64,u64))');
+    expect(fragmentTwoParams.getInputsSighash()).toBe('sum_test(s(u64,u64))');
     let encoded = Interface.getSighash('sum_test(s(u64,u64))');
     expect(hexlify(encoded)).toEqual('0x00000000fd5ec586');
     encoded = Interface.getSighash(fragmentTwoParams);
@@ -125,8 +125,42 @@ describe('Interface', () => {
     expect(decoded.length).toEqual(1);
     expect(decoded[0]).toEqual(42n);
   });
+  it('can calculate the correct sighash for array string values', () => {
+    const fnFragment = FunctionFragment.fromObject({
+      type: 'function',
+      inputs: [
+        {
+          name: 'arg',
+          type: '[s[3]; 3]',
+          components: [
+            {
+              name: '__array_element',
+              type: 's[3]',
+            },
+          ],
+        },
+      ],
+      name: 'takes_array',
+      outputs: [
+        {
+          name: '',
+          type: '[s[3]; 2]',
+          components: [
+            {
+              name: '__array_element',
+              type: 's[3]',
+            },
+          ],
+        },
+      ],
+    });
 
-  it('can calculate the correct sighash for array values', () => {
+    expect(fnFragment.getInputsSighash()).toBe('takes_array(a[s[3];3])');
+    const sighash = Interface.getSighash(fnFragment);
+    expect(hexlify(sighash)).toEqual('0x00000000b80a1c57');
+  });
+
+  it('can calculate the correct sighash for array of u64 values', () => {
     const fnFragment = FunctionFragment.fromObject({
       type: 'function',
       inputs: [
@@ -155,8 +189,46 @@ describe('Interface', () => {
         },
       ],
     });
+
+    expect(fnFragment.getInputsSighash()).toBe('takes_array(a[u16;3])');
     const sighash = Interface.getSighash(fnFragment);
-    expect(hexlify(sighash)).toEqual('0x00000000058734b9');
+    expect(hexlify(sighash)).toEqual('0x00000000101cbeb5');
+  });
+
+  it('can calculate the correct sighash for enum', () => {
+    const fnFragment = FunctionFragment.fromObject({
+      type: 'function',
+      inputs: [
+        {
+          name: 'enum_arg',
+          type: 'enum TestEnum',
+          components: [
+            {
+              name: 'Value',
+              type: 'bool',
+              components: null,
+            },
+            {
+              name: 'Data',
+              type: 'bool',
+              components: null,
+            },
+          ],
+        },
+      ],
+      name: 'take_enum',
+      outputs: [
+        {
+          name: '',
+          type: 'bool',
+          components: null,
+        },
+      ],
+    });
+
+    expect(fnFragment.getInputsSighash()).toBe('take_enum(e(bool,bool))');
+    const sighash = Interface.getSighash(fnFragment);
+    expect(hexlify(sighash)).toEqual('0x00000000424d6522');
   });
 
   it('can encode and decode function data with array values', () => {
@@ -191,7 +263,7 @@ describe('Interface', () => {
       },
     ]);
     expect(hexlify(functionInterface.encodeFunctionData('takes_array', [[1, 2, 3]]))).toEqual(
-      '0x00000000058734b90000000000000001000000000000000100000000000000020000000000000003'
+      '0x00000000101cbeb50000000000000001000000000000000100000000000000020000000000000003'
     );
   });
 
@@ -256,8 +328,208 @@ describe('Interface', () => {
       },
     ]);
 
-    expect(functionInterface.getFunction('entry_one').format()).toEqual(
+    expect(functionInterface.getFunction('entry_one').getInputsSighash()).toEqual(
       'entry_one(u64,s(bool,u64))'
+    );
+  });
+
+  it('can encode array of structs', () => {
+    functionInterface = new Interface([
+      {
+        type: 'function',
+        name: 'entry_one',
+        inputs: [
+          {
+            name: 'arg1',
+            type: '[struct MyStruct; 3]',
+            components: [
+              {
+                name: '__array_element',
+                type: 'struct MyStruct',
+                components: [
+                  {
+                    name: 'bim',
+                    type: 'str[3]',
+                    components: null,
+                    typeArguments: null,
+                  },
+                  {
+                    name: 'bam',
+                    type: 'enum MyEnum',
+                    components: [
+                      {
+                        name: 'Foo',
+                        type: 'u64',
+                        components: null,
+                        typeArguments: null,
+                      },
+                      {
+                        name: 'Bar',
+                        type: 'bool',
+                        components: null,
+                        typeArguments: null,
+                      },
+                      {
+                        name: 'Din',
+                        type: 'bool',
+                        components: null,
+                        typeArguments: null,
+                      },
+                    ],
+                    typeArguments: null,
+                  },
+                ],
+                typeArguments: null,
+              },
+            ],
+            typeArguments: null,
+          },
+        ],
+        outputs: [
+          {
+            name: '',
+            type: 'str[3]',
+            components: null,
+            typeArguments: null,
+          },
+        ],
+      },
+    ]);
+
+    expect(functionInterface.getFunction('entry_one').getInputsSighash()).toEqual(
+      'entry_one(a[s(str[3],e(u64,bool,bool));3])'
+    );
+  });
+
+  it('can encode struct with dynamic typing', () => {
+    functionInterface = new Interface([
+      {
+        type: 'function',
+        name: 'entry_one',
+        inputs: [
+          {
+            name: 'my_u64',
+            type: 'u64',
+          },
+          {
+            name: 'my_struct',
+            type: 'struct MyStruct',
+            components: [
+              {
+                name: 'dummy_a',
+                type: 'bool',
+              },
+              {
+                name: 'dummy_b',
+                type: 'u64',
+              },
+            ],
+            typeArguments: [
+              {
+                name: 'T',
+                type: 'b256',
+              },
+              {
+                name: 'U',
+                type: 'bool',
+              },
+            ],
+          },
+        ],
+        outputs: [{ name: 'ret', type: 'u64' }],
+      },
+    ]);
+
+    expect(functionInterface.getFunction('entry_one').getInputsSighash()).toEqual(
+      'entry_one(u64,s<b256,bool>(bool,u64))'
+    );
+  });
+
+  it('can encode array of structs with dynamic typing', () => {
+    functionInterface = new Interface([
+      {
+        type: 'function',
+        name: 'entry_one',
+        inputs: [
+          {
+            name: 'arg1',
+            type: '[struct MyStruct; 3]',
+            components: [
+              {
+                name: '__array_element',
+                type: 'struct MyStruct',
+                components: [
+                  {
+                    name: 'bim',
+                    type: 'str[3]',
+                    components: null,
+                    typeArguments: null,
+                  },
+                  {
+                    name: 'bam',
+                    type: 'enum MyEnum',
+                    components: [
+                      {
+                        name: 'Foo',
+                        type: 'u64',
+                        components: null,
+                        typeArguments: null,
+                      },
+                      {
+                        name: 'Bar',
+                        type: 'bool',
+                        components: null,
+                        typeArguments: null,
+                      },
+                      {
+                        name: 'Din',
+                        type: 'bool',
+                        components: null,
+                        typeArguments: null,
+                      },
+                    ],
+                    typeArguments: [
+                      {
+                        name: 'V',
+                        type: 'bool',
+                        components: null,
+                        typeArguments: null,
+                      },
+                    ],
+                  },
+                ],
+                typeArguments: [
+                  {
+                    name: 'T',
+                    type: 'str[3]',
+                    components: null,
+                    typeArguments: null,
+                  },
+                  {
+                    name: 'U',
+                    type: 'bool',
+                    components: null,
+                    typeArguments: null,
+                  },
+                ],
+              },
+            ],
+            typeArguments: null,
+          },
+        ],
+        outputs: [
+          {
+            name: '',
+            type: 'str[3]',
+            components: null,
+            typeArguments: null,
+          },
+        ],
+      },
+    ]);
+
+    expect(functionInterface.getFunction('entry_one').getInputsSighash()).toEqual(
+      'entry_one(a[s<str[3],bool>(str[3],e<bool>(u64,bool,bool));3])'
     );
   });
 });