+++ packages/contract/src/__test__/contract.test.ts
@@ -1,4 +1,5 @@
 import { NativeAssetId, ZeroBytes32 } from '@fuel-ts/constants';
+import { multiply } from '@fuel-ts/math';
 import { Provider } from '@fuel-ts/providers';
 import { TestUtils } from '@fuel-ts/wallet';
 
@@ -34,6 +35,8 @@ const complexFragment = {
   outputs: [],
   type: 'function',
 };
+
+const AltToken = '0x0101010101010101010101010101010101010101010101010101010101010101';
 
 describe('Contract', () => {
   it('generates function methods on a simple contract', async () => {
@@ -233,10 +236,10 @@ describe('Contract', () => {
           forward: [100, NativeAssetId],
         }),
         contract.functions.return_context_amount().callParams({
-          forward: [200, '0x0101010101010101010101010101010101010101010101010101010101010101'],
+          forward: [200, AltToken],
         }),
         contract.functions.return_context_asset().callParams({
-          forward: [0, '0x0101010101010101010101010101010101010101010101010101010101010101'],
+          forward: [0, AltToken],
         }),
       ])
       .txParams({
@@ -245,11 +248,7 @@ describe('Contract', () => {
         gasLimit: 2000000,
       })
       .call<[bigint, bigint, string]>();
-    expect(value).toEqual([
-      100n,
-      200n,
-      '0x0101010101010101010101010101010101010101010101010101010101010101',
-    ]);
+    expect(value).toEqual([100n, 200n, AltToken]);
   });
 
   it('Check if gas per call is lower than transaction', async () => {
@@ -263,7 +262,7 @@ describe('Contract', () => {
             gasLimit: 100,
           }),
           contract.functions.return_context_amount().callParams({
-            forward: [200, '0x0101010101010101010101010101010101010101010101010101010101010101'],
+            forward: [200, AltToken],
             gasLimit: 200,
           }),
         ])
@@ -307,5 +306,210 @@ describe('Contract', () => {
 
     expect(Number(value[1])).toBeGreaterThanOrEqual(1_000_000 * allowedError);
     expect(Number(value[1])).toBeLessThanOrEqual(1_000_000);
+  });
+
+  it('Get transaction cost', async () => {
+    const contract = await setup();
+
+    const invocationScope = contract.multiCall([
+      contract.functions.return_context_amount().callParams({
+        forward: [100, NativeAssetId],
+      }),
+      contract.functions.return_context_amount().callParams({
+        forward: [200, AltToken],
+      }),
+    ]);
+    const transactionCost = await invocationScope.getTransactionCost();
+
+    expect(transactionCost.bytePrice).toBe(0n);
+    expect(transactionCost.gasPrice).toBe(0n);
+    expect(transactionCost.fee).toBeGreaterThanOrEqual(0n);
+    expect(transactionCost.gasUsed).toBeGreaterThan(1000n);
+
+    const { value } = await invocationScope
+      .txParams({
+        bytePrice: transactionCost.bytePrice,
+        gasPrice: transactionCost.gasPrice,
+        gasLimit: transactionCost.gasUsed,
+      })
+      .call<[bigint, bigint]>();
+
+    expect(value).toEqual([100n, 200n]);
+  });
+
+  it('Get transaction cost with bytePrice and gasPrice 1', async () => {
+    const contract = await setup();
+
+    const invocationScope = contract
+      .multiCall([
+        contract.functions.return_context_amount().callParams({
+          forward: [100, NativeAssetId],
+        }),
+        contract.functions.return_context_amount().callParams({
+          forward: [200, AltToken],
+        }),
+      ])
+      .txParams({
+        gasPrice: 1,
+        bytePrice: 1,
+      });
+    // Get transaction cost using bytePrice and gasPrice from
+    // invocation scope
+    const transactionCost = await invocationScope.getTransactionCost();
+
+    expect(transactionCost.bytePrice).toBe(1n);
+    expect(transactionCost.gasPrice).toBe(1n);
+    expect(transactionCost.fee).toBeGreaterThanOrEqual(2n);
+    expect(transactionCost.gasUsed).toBeGreaterThan(1000n);
+
+    // Test that gasUsed is correctly calculated
+    // and can be used as gasLimit
+    const { value } = await invocationScope
+      .txParams({
+        bytePrice: transactionCost.bytePrice,
+        gasPrice: transactionCost.gasPrice,
+        gasLimit: transactionCost.gasUsed,
+      })
+      .call<[bigint, bigint]>();
+
+    expect(value).toEqual([100n, 200n]);
+  });
+
+  it('Get transaction cost with bytePrice and gasPrice 2', async () => {
+    const contract = await setup();
+
+    const invocationScope = contract.multiCall([
+      contract.functions.return_context_amount().callParams({
+        forward: [100, NativeAssetId],
+      }),
+      contract.functions.return_context_amount().callParams({
+        forward: [200, AltToken],
+      }),
+    ]);
+    // Get transaction cost using bytePrice and gasPrice
+    // override by SDK user
+    const transactionCost = await invocationScope.getTransactionCost({
+      gasPrice: 2,
+      bytePrice: 2,
+    });
+
+    expect(transactionCost.bytePrice).toBe(2n);
+    expect(transactionCost.gasPrice).toBe(2n);
+    expect(transactionCost.fee).toBeGreaterThanOrEqual(4n);
+    expect(transactionCost.gasUsed).toBeGreaterThan(1000n);
+
+    // Test that gasUsed is correctly calculated
+    // and can be used as gasLimit
+    const { value } = await invocationScope
+      .txParams({
+        bytePrice: transactionCost.bytePrice,
+        gasPrice: transactionCost.gasPrice,
+        gasLimit: transactionCost.gasUsed,
+      })
+      .call<[bigint, bigint]>();
+
+    expect(value).toEqual([100n, 200n]);
+  });
+
+  it('Fail before submit if gasLimit is lower than gasUsed', async () => {
+    const contract = await setup();
+
+    const invocationScope = contract.functions.return_context_amount().callParams({
+      forward: [100, NativeAssetId],
+    });
+    const { gasUsed } = await invocationScope.getTransactionCost({
+      tolerance: 0,
+    });
+
+    const gasLimit = multiply(gasUsed, 0.5);
+    await expect(async () => {
+      await invocationScope
+        .txParams({
+          gasLimit,
+        })
+        .call<bigint>();
+    }).rejects.toThrowError(`gasLimit(${gasLimit}) is lower than the required (${gasUsed})`);
+  });
+
+  it('calls array functions', async () => {
+    const contract = await setup();
+
+    const { value: arrayBoolean } = await contract.functions
+      .take_array_boolean([true, false, false])
+      .call();
+
+    expect(arrayBoolean).toEqual(true);
+
+    const { value: arrayNumber } = await contract.functions.take_array_number([1, 2, 3]).call();
+
+    expect(arrayNumber).toEqual(1n);
+
+    const { value: arrayReturnShuffle } = await contract.functions
+      .take_array_string_shuffle(['abc', 'efg', 'hij'])
+      .call();
+
+    expect(arrayReturnShuffle).toEqual(['hij', 'abc', 'efg']);
+
+    const { value: arrayReturnSingle } = await contract.functions
+      .take_array_string_return_single(['abc', 'efg', 'hij'])
+      .call();
+
+    expect(arrayReturnSingle).toEqual(['abc']);
+  });
+
+  it('calls enum functions', async () => {
+    const contract = await setup();
+
+    const { value: enumB256ReturnValue } = await contract.functions
+      .take_b256_enum({
+        Value: '0xd5579c46dfcc7f18207013e65b44e4cb4e2c2298f4ac457ba8f82743f31e930b',
+      })
+      .call();
+
+    expect(enumB256ReturnValue).toEqual(
+      '0xd5579c46dfcc7f18207013e65b44e4cb4e2c2298f4ac457ba8f82743f31e930b'
+    );
+
+    const { value: enumB256ReturnData } = await contract.functions
+      .take_b256_enum({
+        Data: '0x1111111111111111111111111111111111111111111111111111111111111111',
+      })
+      .call();
+
+    expect(enumB256ReturnData).toEqual(
+      '0x1111111111111111111111111111111111111111111111111111111111111111'
+    );
+
+    const { value: enumBoolReturnValue } = await contract.functions
+      .take_bool_enum({
+        Value: true,
+      })
+      .call();
+
+    expect(enumBoolReturnValue).toEqual(true);
+
+    const { value: enumBoolReturnData } = await contract.functions
+      .take_bool_enum({
+        Data: false,
+      })
+      .call();
+
+    expect(enumBoolReturnData).toEqual(false);
+
+    const { value: enumStrReturnValue } = await contract.functions
+      .take_string_enum({
+        Value: 'abc',
+      })
+      .call();
+
+    expect(enumStrReturnValue).toEqual('abc');
+
+    const { value: enumStrReturnData } = await contract.functions
+      .take_string_enum({
+        Data: 'efg',
+      })
+      .call();
+
+    expect(enumStrReturnData).toEqual('efg');
   });
 });