+++ packages/predicate/src/predicate.ts
@@ -12,6 +12,7 @@ import type {
   Coin,
 } from '@fuel-ts/providers';
 import { ScriptTransactionRequest } from '@fuel-ts/providers';
+import { MAX_GAS_PER_TX } from '@fuel-ts/transactions';
 import type { Wallet } from '@fuel-ts/wallet';
 
 type BuildPredicateOptions = {
@@ -44,7 +45,7 @@ export class Predicate {
       ...predicateOptions,
     };
     const request = new ScriptTransactionRequest({
-      gasLimit: 1000000,
+      gasLimit: MAX_GAS_PER_TX,
       ...options,
     });
 
@@ -53,8 +54,7 @@ export class Predicate {
 
     const requiredCoinQuantities: CoinQuantityLike[] = [];
     if (options.fundTransaction) {
-      const amount = request.calculateFee();
-      requiredCoinQuantities.push([amount]);
+      requiredCoinQuantities.push(request.calculateFee());
     }
 
     if (requiredCoinQuantities.length) {
@@ -98,7 +98,7 @@ export class Predicate {
       ...predicateOptions,
     };
     const request = new ScriptTransactionRequest({
-      gasLimit: 1000000,
+      gasLimit: MAX_GAS_PER_TX,
       ...options,
     });
 
@@ -124,8 +124,7 @@ export class Predicate {
 
     const requiredCoinQuantities: CoinQuantityLike[] = [];
     if (options.fundTransaction) {
-      const amount = request.calculateFee();
-      requiredCoinQuantities.push([amount]);
+      requiredCoinQuantities.push(request.calculateFee());
     }
 
     if (requiredCoinQuantities.length) {