+++ packages/mnemonic/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/keystore@0.10.0
+  - @fuel-ts/wordlists@0.10.0
+
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/keystore@0.9.0
+  - @fuel-ts/wordlists@0.9.0
+
 ## 0.8.0
 
 ### Minor Changes